Configure linting
Since we're starting from scratch now is a great time to be particularly
harsh around linting.  This throws errors when linting warnings appear,
it also prevents an html report, opting for command line output instead.

The decision to treat warnings as errors may prove cumbersome,
especially because dependency upgrades are errors (so we may be working
on an unrelated issue and need to upgrade a dependency before we can
merge). If this becomes too obnoxious we can turn it off at a later
date.

Issue #4 Set up core project + documentation