Add introduction page
This creates a very simple introduction page, which is what users will
see when opening the application.  It features the title, a "tutorial"
image (TBD), and a button to move to the next page.

Since the template we used has a starting page with a button, this
re-uses (re-names) the "first fragment" that was generated by the
template.

You'll notice that we're using the strings.xml for all text-based
aspects of the code.  This is going to be a best practice that will
allow us to support multiple languages more easily some day.

For now we aren't worrying about look and feel / styles at all.  This is
purely about creating the UI elements.

Issue #15 Create the opening frame